Add skips and make code run
