+++ src/Stratis.Bitcoin.FullNode.sln
@@ -158,6 +158,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Stratis.Features.FederatedP
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Stratis.Features.FederatedPeg.Tests", "Stratis.Features.FederatedPeg.Tests\Stratis.Features.FederatedPeg.Tests.csproj", "{DE629FE5-BC8A-4597-B19C-DDAB5EC1528D}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Stratis.Features.FederatedPeg.IntegrationTests", "Stratis.Features.FederatedPeg.IntegrationTests\Stratis.Features.FederatedPeg.IntegrationTests.csproj", "{D71A7BBE-E64A-4F2E-B2A9-6D742FE94ED9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -408,6 +410,10 @@ Global
 		{DE629FE5-BC8A-4597-B19C-DDAB5EC1528D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DE629FE5-BC8A-4597-B19C-DDAB5EC1528D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DE629FE5-BC8A-4597-B19C-DDAB5EC1528D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D71A7BBE-E64A-4F2E-B2A9-6D742FE94ED9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D71A7BBE-E64A-4F2E-B2A9-6D742FE94ED9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D71A7BBE-E64A-4F2E-B2A9-6D742FE94ED9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D71A7BBE-E64A-4F2E-B2A9-6D742FE94ED9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -466,6 +472,7 @@ Global
 		{398E3C08-1628-4985-B3F3-36396A276E6F} = {1B9A916F-DDAC-4675-B424-EDEDC1A58231}
 		{C2781309-00DE-4AC4-ABBB-12FF26B8BFB7} = {15D29FFD-6142-4DC5-AFFD-10BA0CA55C45}
 		{DE629FE5-BC8A-4597-B19C-DDAB5EC1528D} = {EAE139C2-B19C-4905-9117-8A4068ABD3D2}
+		{D71A7BBE-E64A-4F2E-B2A9-6D742FE94ED9} = {EAE139C2-B19C-4905-9117-8A4068ABD3D2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6C780ABA-5872-4B83-AD3F-A5BD423AD907}
