+++ src/Stratis.Features.FederatedPeg.IntegrationTests/Utils/SidechainFederationNodeRunner.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using NBitcoin;
+﻿using NBitcoin;
 using Stratis.Bitcoin;
 using Stratis.Bitcoin.Builder;
 using Stratis.Bitcoin.Configuration;
@@ -11,40 +8,49 @@ using Stratis.Bitcoin.Features.MemoryPool;
 using Stratis.Bitcoin.Features.Notifications;
 using Stratis.Bitcoin.Features.RPC;
 using Stratis.Bitcoin.Features.SmartContracts;
-using Stratis.Bitcoin.Features.SmartContracts.PoA;
 using Stratis.Bitcoin.Features.SmartContracts.Wallet;
-using Stratis.Bitcoin.Features.Wallet;
 using Stratis.Bitcoin.IntegrationTests.Common;
 using Stratis.Bitcoin.IntegrationTests.Common.Runners;
-using Stratis.FederatedPeg.Features.FederationGateway;
 
-namespace Stratis.FederatedPeg.IntegrationTests.Utils
+namespace Stratis.Features.FederatedPeg.IntegrationTests.Utils
 {
-    public class SidechainUserNodeRunner : NodeRunner
+    public class SidechainFederationNodeRunner : NodeRunner
     {
-        public SidechainUserNodeRunner(string dataDir, string agent, Network network)
+        private bool testingFederation;
+
+        public SidechainFederationNodeRunner(string dataDir, string agent, Network network, bool testingFederation)
             : base(dataDir, agent)
         {
             this.Network = network;
+
+            this.testingFederation = testingFederation;
         }
 
         public override void BuildNode()
         {
             var settings = new NodeSettings(this.Network, args: new string[] { "-conf=poa.conf", "-datadir=" + this.DataFolder });
 
-            this.FullNode = (FullNode)new FullNodeBuilder()
+            var builder = new FullNodeBuilder()
                 .UseNodeSettings(settings)
                 .UseBlockStore()
                 .AddSmartContracts()
-                .UseSmartContractPoAConsensus()
-                .UseSmartContractPoAMining()
                 .UseSmartContractWallet()
                 .UseReflectionExecutor()
+                .AddFederationGateway()
+                .UseFederatedPegPoAMining()
                 .UseMempool()
+                .UseTransactionNotification()
+                .UseBlockNotification()
                 .UseApi()
-                .MockIBD()
                 .AddRPC()
-                .Build();
+                .MockIBD();
+
+            if (!this.testingFederation)
+            {
+                builder.UseTestFedPegBlockDefinition();
+            }
+
+            this.FullNode = (FullNode) builder.Build();
         }
     }
 }
