+++ src/Stratis.Features.FederatedPeg.IntegrationTests/Utils/SidechainUserNodeRunner.cs
@@ -5,18 +5,18 @@ using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Features.Api;
 using Stratis.Bitcoin.Features.BlockStore;
 using Stratis.Bitcoin.Features.MemoryPool;
-using Stratis.Bitcoin.Features.Notifications;
 using Stratis.Bitcoin.Features.RPC;
 using Stratis.Bitcoin.Features.SmartContracts;
+using Stratis.Bitcoin.Features.SmartContracts.PoA;
 using Stratis.Bitcoin.Features.SmartContracts.Wallet;
+using Stratis.Bitcoin.IntegrationTests.Common;
 using Stratis.Bitcoin.IntegrationTests.Common.Runners;
-using Stratis.FederatedPeg.Features.FederationGateway;
 
-namespace Stratis.FederatedPeg.IntegrationTests.Utils
+namespace Stratis.Features.FederatedPeg.IntegrationTests.Utils
 {
-    public class SidechainWithSmartContractsNodeRunner : NodeRunner
+    public class SidechainUserNodeRunner : NodeRunner
     {
-        public SidechainWithSmartContractsNodeRunner(string dataDir, string agent, Network network)
+        public SidechainUserNodeRunner(string dataDir, string agent, Network network)
             : base(dataDir, agent)
         {
             this.Network = network;
@@ -30,14 +30,13 @@ namespace Stratis.FederatedPeg.IntegrationTests.Utils
                 .UseNodeSettings(settings)
                 .UseBlockStore()
                 .AddSmartContracts()
+                .UseSmartContractPoAConsensus()
+                .UseSmartContractPoAMining()
                 .UseSmartContractWallet()
                 .UseReflectionExecutor()
-                .AddFederationGateway()
-                .UseFederatedPegPoAMining()
                 .UseMempool()
-                .UseTransactionNotification()
-                .UseBlockNotification()
                 .UseApi()
+                .MockIBD()
                 .AddRPC()
                 .Build();
         }
