+++ src/Stratis.Features.FederatedPeg.IntegrationTests/xunit.runner.json
